Fixed details element for browsers with native support
The details > * {display:none;} CSS rule was hiding the <details>
content in Chrome since it natively supports the element.

Fix was to add a details:not(.polyfill) rule that sets display to
block.  Using the :not() selector reduces the flash of unstyled
content (FOUC) seen in browsers that use the details polyfill.