Add Offsets, Generate Grids by Sass function
I've added Offsets to the grid system so we don't have
to fill up our websites with empty divs that contain spans
for spacing.

Spans and Offsets are generated using SASS functions to
cut down on the amount of code in the SASS Files and make it
more readable for developers.

We lose the ability to use @extend but that is not as important
as we minimize the css anyways.