Column count CSS changed to match naming conventions
column-two, column-three and column-four change to column-2, column-3
and column-4 to match CSS naming conventions.  As this is not a popular
feature and only works in IE9+, there isn't much uptake on it yet so old
CSS names removed.   No legacy support provided as it's new as of WET 3
anyhow.