Fixes for when link is both a download and external
Since only one background image can be displayed with CSS, the
padding-left was still present when the file was a download but the
image wouldn't come through because it was also external
(rel="external") so the padding left is not always "0" on rel external
to make sure there is no dead space.

Also added the external icon example to the master list of icons.