Temporary work around for libsass silent error on various color functions
Libsass on a windows platform is generating silent errors on several
sass color functions ( rgba, rgb, hsla,hsl ). It appears that the c++
function is getting a bit confused and trying to interpolate the
variables and failing. The proposed hack will allow our build to run,
since it wraps rgba(....) in an unquote function that sass then
correctly treats as a literal in its interpolation instead of handing
off to the broken function.

There will be a issue logged with node-sass and libsass to look into
this, but there is a general feeling that it will be address at a later
date, since this behaviour is not experienced in Linux and Mac variants.