Fix because DOM spec is dumb
It's so stupid that it's perfectly valid to use
boolean attributes in HTML5 (e.g. open) but that
the DOM spec requires setting a value in the
Element.setAttribute method (e.g. open="open").

Morons.

https://developer.mozilla.org/en-US/docs/Web/API/Element.setAttribute

BTW, this fixes the previous issue where some
of the details/summary instances did not always
set their open/closed state properly on load.