Multimedia player: Added support for youtu.be URLs.
The multimedia player's YouTube variant used to only extract video IDs by checking the v parameter in the video URL's query string. But that didn't lend itself to shortened YouTube URLs (e.g. https://youtu.be/9znKJqnFuuY).

With this change, if a video URL doesn't contain a v parameter, the player will strip the first forward slash from the URL's pathname and treat the result as the video's ID.