Proposed solution for "Twitter feed not loading in IE"
For those with AEM or another CMS that does not let you edit the code, they could simply update their wet-beow.js with that new Twitter.js version.

Note: Since our current website does not have WET 4 and that we are not on Canada.ca, the language part starting at "var pathName" has not been tested but should be working. Please adapt the condition (if, else if) as needed to the URL that you have on your website.