Form validation: Reduce aggressiveness of onfocusout validation
This partially-reverts #7921.

One of its changes caused form validation to always get triggered when focusing out of fields. The end result was that required field validation became drastically more aggressive, arguably to the point of impeding usability.

For example:
* Tabbing through required inputs, checkboxes or radio buttons that had never received any user input would immediately trigger required field errors
* Ticking and then unticking any required checkbox would immediately trigger a required field error

Examples of situations where that would annoy users:
* Keyboard user getting bombarded with required field errors while tabbing through a form
* Mouse/Touch screen user accidentally clicking or touching a random required field, then having to deal with fields visually-shifting downwards

According to the jQuery Validation Plugin's reference documentation (https://jqueryvalidation.org/reference/#link-validation-event), the plugin is meant to provide quick contextual error feedback, without going so far as to annoy users. It considers displaying error messages without having received user input to cross that line.

I personally find the behaviour I'm reverting to be really annoying and aren't the only one that dislikes it... That behaviour was also brought up as a concern by a usability subject matter expert that reviewed a form project I'm currently helping out with.

Fixes #8353 and partially-addresses #8688.

Related to #8111, #8160 and #8066.