Fix clicking on child elements of anchor
If tabs are used in a modal window and the tab summary element has child elements (span, abbr, etc...) then clicks to the tab anchor (which is generated) throw an error on line 328. The error is thrown because the event assumes that event.target is an anchor. Since it is not, it has no href attribute and the code fails. By changing to event.currentTarget, it ensures that the eventTarget element is the anchor from the event filter.