Bootstrap overrides: Darken panel borders
This increases the contrast between panel headings and borders to >= 3:1. The resulting effect looks very stark.

The upside of this change is that it makes the shapes of panels more perceptible to users with poor vision. It may also benefit WCAG 2.1 success criterion 1.4.11 (Non-text Contrast) - especially when panels are used with stretched links.