Multimedia player: Fix AJAX race condition in Firefox
Players brought in via the Data Ajax plugin sometimes tried initializing themselves twice in Firefox.

When it happened, the first attempt occurred before the AJAXed-in player had received a "wb-auto-*" ID - resulting in the player and media element's IDs getting set to "undefined" and "undefined-md" respectively.

The second attempt then treated the player's transformed HTML markup as if it was the initial markup. That caused several issues, such as a "video" class getting added to audio players (producing a large play icon overlapping the player controls), ID mismatches and a JavaScript console error related to the media variable being undefined.

This resolves it via the following changes:
* Add a check to ensure the media variable is truthy before proceeding with initialization
* Move most variable declarations after the media check
* Get an ID from wb.getId() if the player doesn't already have one
  * Note: Getting an ID at this point will prevent the second initialization from occurring. But the media check is still necessary for players that use hardcoded IDs.
* Don't check media.currentSrc in the condition that triggers renderUIEvent (speeds things up by triggering it on the first init instead of the second, higher-level media check makes this one redundant)