Equal heights: Vary example 1's row heights
The demo page's first example demonstrates the plugin's basic concept, but re-uses the exact same content in all three rows. In large view, that gives off the impression that the plugin is unable to distinguish "virtual" rows.

In practice, the plugin is sophisticated-enough to automatically detect new rows and keep track of varying row heights. So why not show it off?

This changes the first example as follows to better demonstrate that capability:
* Place all rows in the same wb-eqht container
* Make the first row's paragraphs 4 lines apiece
* Halve the amount of content lines in each subsequent row in large view
* Use unique headings for each row
* Note: Shortened French content to get the same number of content lines as the English version in large view

The first two examples in the "pure CSS" equal heights demo have also been updated to match:
* Same changes as the normal equal heights demo
* Reduce demo's custom CSS padding from 15px to 10px (to prevent the French examples' H3 headings from spanning across multiple lines in large view)