Update _core.scss
Change bootstrap width default of <dt> for horizontal from 160px to 20ch in order to allow both WET4 and GCWeb to allow the same amount of characters in the <dt> to remain the same for both themes without any overrides.  Added a new class, .dt-max to allow the longest word in a set of  <dt> to dictate the width without any word break.