Twitter embedded timeline: Add invisible start notice
Works much better in screen readers.

Previously, the skip to start link pointed to the timeline iframe's parent div... which caused a ~30-50 second period of silence in NVDA, followed by the entire timeline getting automatically announced.

This addresses it by adding an invisible start notice just before the timeline and changing the skip to start link's destination to go there instead. That causes screen readers to only announce the start notice at first - without delay.

A "fake" outline is also added to the timeline for sighted users (to retain the previous behaviour's visual behaviour - which works better for that audience). The start notice is never visually-revealed to them (showing it as an overlay would've covered part of the timeline... not ideal).