Twitter embedded timeline: Improve handling of missing i18n vars
#9704's changes had the potential of going into an infinite loop if one of its i18n variables was undefined.

The infinite loop tied into an if condition in the MutationObserver:
* Its purpose is to replace the Twitter widget's English-only iframe title with an i18n version from WET
* It checks for attribute mutations on iframe elements and whether the iframe's title attribute matches the i18nText.timelineTitle variable
* If the current iframe title isn't strictly equal to the i18n variable, the former gets set to the latter (triggering another attribute mutation)

Normally, the if condition should become truthy in the subsequent attribute mutation.

When the i18n variable wasn't set, this never happened. Why? Because missing i18n variables are treated as undefined. So the iframe title got set to "undefined" (implicit conversion of undefined into a string)... Then the next mutation's if condition didn't match ("undefined" !== undefined). So the if condition attempted to set the iframe title to "undefined" once again - triggering yet another attribute mutation. This sequence of events repeated forever.

This fixes it by checking whether the i18n variable is a string before changing the iframe's title attribute.

Related changes:
* Show a console warning if the plugin's i18n variables appear to be missing
* Don't add skip links if one of its i18n variables isn't a string (prevents console errors, no point in adding the links without content)