Focus: Assign focus more resiliently
If this plugin is used to focus onto a non-interactive element, it adds a tabindex="-1" attribute (to make the destination focusable) before proceeding.

Since adding tabindex="-1" isn't always necessary, the plugin exempts elements that already have a tabindex attribute, as well as "known" interactive elements.

However...
* The plugin removed interactive elements it wasn't aware of (like summary) from sequential keyboard navigation:
  * End result was WCAG 2.x failures: SC 2.1.1 and 2.4.3 (via F44)
* Tying exemptions to a list of "known" interactive elements is a bad practice:
  * Isn't future-proof
  * Keeping it up-to-date is a maintenance burden
  * Interactive content as defined by the HTML spec (https://html.spec.whatwg.org/multipage/dom.html#interactive-content) isn't suitable for that purpose (e.g. details and label are technically interactive... but aren't keyboard-tabbable)
  * Doesn't account for different behaviours across browser rendering engines (e.g. video element without the controls attribute is tabbable in Gecko, but not in Blink)

This alleviates the need for an exemptions list by revamping how focusing works:
1. The plugin will first try to focus directly onto the element
2. If the element hasn't gained focus, the plugin will add a tabindex="-1" attribute and retry
3. If the element still hasn't gained focus... the plugin will remove tabindex="-1" and show a console error

Credit goes to @fsnoddy for initially spotting this issue (by using the toggle plugin's {"print": "on"} configuration option on a summary element). Also added "boring" print state examples to the toggle plugin's demo page to show it in action.