Facebook embedded pages: Add overflow-x scrollbar
Facebook's page widget isn't fully responsive:
* Default width is 340px
* Officially supports widths between 180-500px (doesn't support percentages)
* Automatically shrinks to fit within its container when inited (via data-adapt-container-width="true", default behaviour)
* Can't shrink or expand after initing (can't use CSS to resize its iframe either... width is set inside the iframe's body)

As a result, if the widget initially renders itself within a wide container, resizing the viewport won't cause it to scale... So if the viewport width shrinks and the widget overflows beyond its container, the entire page will become horizontally-scrollable.

This limits the damage by adding an overflow-x property to the widget's container (to make only the widget itself scroll horizontally).