Update _base.scss
xplanation of the change/update:
Change was to remove the legacy CSS code for details/summary for IE11 (which did not support this element) as its no longer required.  The removal was for the display:none and visibility:  visible/hidden styles.

This fixes two issues: 
1. It prevents the focus of the keyboard tab from moving to a hidden tab in desktop if it contains an details element thus violating WCAG SC 2.4.7 (Level AA) Focus Visible.  
2. It allows the user to use Ctrl+F5 to search for a word/phrase inside of a details element and also opens the element if the word/phrase exists.  

Impact:
The only impact this will have may be for anybody with an unsupported browser that does not support detail elements (Internet Explorer), or very early browser versions.  If there is a desire to support this, a script could be written or a fallback stylesheet could be created.  Otherwise the issue should be minimal.