Content - Form validation (docs): Add "required" class to label code sample (#9867)
The documentation's "How to implement" section's second step contains a code sample depicting a required label... without a "required" class on the label element.

That appears to be a mistake. It's extremely unlikely any implementers would deliberately omit the "required" class in practice given that it's needed to set the bolded "(required)" text's colour to red. The class is also used across-the-board in all of the demo page's required examples.

This resolves it by adding the missing "required" class to the documentation's required label code sample.