Lightbox: Close modal popups via Escape key
The lightbox plugin's documentation states that it implements the WAI-ARIA Dialog (Modal) design pattern.

The pattern's keyboard interaction documentation states that the Escape key is expected to be capable of closing the dialog. Same for Alert and Message Dialogs (stricter pattern that seems more in line with how WET's modal popups actually work).

Most of the lightbox demo page's examples can currently be closed via the Escape key, but the modal examples can't. Why? Because the lightbox plugin leverages Magnific Popup's "modal" option (https://dimsemenov.com/plugins/magnific-popup/documentation.html#modal) - which disables the Escape key.

This restores the Escape key's expected behaviour in order to adhere to the pattern. It works by triggering a "fake" click event on the modal's last close button.

Related change:
* Session timeout (confirm method): Use setTimeout() when changing window.location.href to support lightbox Escape key presses

Fixes #9134. Related to #7919.

PS:
The link to the pattern's documentation is currently broken, but can be found at the following locations:
* Wayback Machine snapshots:
  * 2014 (as it stood when the link was first added to the lightbox documentation):
    * https://web.archive.org/web/20140805134917/https://www.w3.org/TR/wai-aria-practices/#dialog_modal
  * 2022 (just before being replaced with a move notice):
    * https://web.archive.org/web/20220516223228/https://www.w3.org/TR/wai-aria-practices/#dialog_modal
* ARIA Authoring Practices Guide (APG):
  * Dialog (Modal) Pattern:
    * https://www.w3.org/WAI/ARIA/apg/patterns/dialog-modal/
  * Alert and Message Dialogs Pattern:
    * https://www.w3.org/WAI/ARIA/apg/patterns/alertdialog/