Patch - Steps form (quiz): Remove "phantom" spacing below legends (#10012)
To disable the "magic" layout of fieldset legends, WET floats legends to the left by default.

The steps form plugin's quiz option previously disabled legend floating to prevent layout issues with nested fieldsets. But that in turn re-activated the legend's "magic" layout and caused an extra 10px of "phantom" space to appear below the current step's legend.

This resolves it by simplifying the layout fix for nested fieldsets:
* Remove the legend's float: none; and border/padding-top overrides
* Add clear: left; to any elements adjacent to legends to suppress unusual floating behaviour